Add dumb elephant-style arbitrary xonk import
If the search field contains something starting by "https://", then we
consider that it is a xonk that needs to be imported so you can interact
with it. This replaces upstream honk "xzone".

Beware if you are not using my fork but want the feature, `ximport`
has been modified to accept a search query instead of form values.