Remove pleroma.css, add body background image for guests
This background image is only added when you are not logged in; I do not
like having it full time.

You need to place a JPEG file in data/views/background.jpg for it to
work. If there is none, the style will fall back to the default
background color.