WEB: add "/searchxonkers?q=something"
This returns a list of mentionable xonkers given a `q` parameter, and
returns an empty string and a 204 HTTP status if there is no match,
otherwise a JSON string containing all matches for the given `q` in the
`handle@domain` format.

The greedy search is done against the xonkers XID's (URL) last part of
the URL. It is not advised to query less than 3 characters or it may be
slow.

This may be removed if I have no frontend use for it, and until another
commit follow this one I will rewrite git history.