Add mention completion
Add a mention "subform" allowing to search for xonkers and pick one that
will be added to the honk. You can repeat the operation at will.

See 9cd690a38506504f3c022d008c462173e76a1705 that explains what can
really be searched and how.

It is not as practical as a bloated jquery solution, but it is already
way more practical than nothing. To avoid multiple queries with the same
content, a debounce has been set up, so it is a tad more laggy than it
really is.

It will not work with Firefox for Android, see:
https://bugzilla.mozilla.org/show_bug.cgi?id=1535985