Bring back DZ display the simple way
Simplified version of Umio-Yasuno/honk-fisherman@1282db4

Instead of using <details> all the time, only use that when there a
content warning text. This simplifies the DOM.

DZ are open by default, you will need to remove the `open` string in
that diff's <details> tag to hide them if you want to. You need to do
that because it is an hfcs.1 option, and this fork has no hfcs.