Fix the PWA load after the CSP change
I messed up and placed the worker service at the wrong place. Load the
PWA manifest and the service worker only when you are logged in; this
does not make much sense for guests, and this avoids to serve another
javascript file, as `honkpage.js` is served only when logged in, and
would cause performance penalties.