Hide "Add Honker" button if honker is already registered.
This impacts all `/u/somehandle` and `/h?xid=...` pages. Note that
unsubed honkers pages will not have the button as well to resub; there
are no getters for that in database.go, and adding that and the cache
would represent too much overhead just for this usage in my opinion.